refactor(benchmark): move top_k_varlen to its own routine module
top_k_varlen is a sparse-attention KV-index selection primitive, not a
vocabulary-sampling op, so it no longer lives under routines/sampling.py:

- New routines/topk_varlen.py (parse_topk_varlen_args + run_topk_varlen_test)
- Own benchmark_apis category + flashinfer_benchmark.py dispatch
- Drop the "cuda" backend sentinel; --backends defaults to the cc-registry
  union (radix / gvr / radix_cutlass), narrowed per-GPU by the cc-filter
- Rename --vocab_size -> --max_seq_len (arg, variable, and CSV column)
- Remove testTopKVarlen and its top_k_varlen-specific --backends from
  sampling.py (reverted to the plain "cuda" default)
